Add webdriver CI for Linux
